♻️ Refactor genomic variant import notifications out of main NotificationsController file
