+++ ui/src/components/admin/Notifications/NotificationsController.js
@@ -1,6 +1,4 @@
-import React, { useContext, useState } from 'react';
-
-import NOTIFICATION_TYPES from './NotificationTypes';
+import React, { useState } from 'react';
 
 export const NotificationsContext = React.createContext();
 
@@ -60,70 +58,6 @@ const NotificationsProvider = ({ children }) => {
       {children}
     </NotificationsContext.Provider>
   );
-};
-
-export const useGenomicVariantImportNotifications = () => {
-  const {
-    appendNotification,
-    clearNotification,
-    importNotifications,
-    setImportNotifications,
-  } = useContext(NotificationsContext);
-
-  const getImportNotifications = () => {
-    if (!importNotifications) return [];
-    return importNotifications;
-  };
-
-  const addImportNotification = async modelName => {
-    const existingNotification = getImportNotifications().find(x => x.modelName === modelName);
-
-    if (!existingNotification) {
-      const notification = await appendNotification({
-        type: NOTIFICATION_TYPES.LOADING,
-        message: `Importing: Research Variants for ${modelName} are currently importing.`,
-        details:
-          'You can continue to use the CMS, and will be notified when the import is complete.',
-        timeout: false,
-      });
-
-      setImportNotifications(notifications => [
-        ...notifications,
-        {
-          modelName: modelName,
-          notificationId: notification.id,
-        },
-      ]);
-    }
-  };
-
-  const removeImportNotification = modelName => {
-    const existingNotification = getImportNotifications().find(x => x.modelName === modelName);
-
-    if (existingNotification) {
-      clearNotification(existingNotification.notificationId);
-      setImportNotifications(notifications =>
-        notifications.filter(notification => notification.modelName !== modelName),
-      );
-    }
-  };
-
-  const showSuccessfulImportNotification = modelName => {
-    appendNotification({
-      type: NOTIFICATION_TYPES.SUCCESS,
-      message: `Import Successful: Research Somatic Variants for ${modelName} have successfully imported, however are not yet published.`,
-      link: `/admin/model/${modelName}`,
-      linkText: 'View on model page to publish these variants.',
-    });
-  };
-
-  return {
-    importNotifications: getImportNotifications(),
-    setImportNotifications,
-    addImportNotification,
-    removeImportNotification,
-    showSuccessfulImportNotification,
-  };
 };
 
 export default NotificationsProvider;
