+++ ui/src/components/admin/Notifications/index.js
@@ -1,7 +1,6 @@
 export { default as NotificationToaster } from './NotificationToaster';
-export {
-  default as NotificationsProvider,
-  NotificationsContext,
-  useGenomicVariantImportNotifications,
-} from './NotificationsController';
+export { default as NotificationsProvider, NotificationsContext } from './NotificationsController';
 export { default as NOTIFICATION_TYPES } from './NotificationTypes';
+export {
+  default as useGenomicVariantImportNotifications,
+} from './GenomicVariantImportNotifications';
